requirements.prod.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.5
django-rest-knox==4.2.0
drf-yasg==1.21.7
google-api-python-client==2.126.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.67.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
whitenoise==6.6.0
setuptools>=70,<81
wheel
dj-database-url==2.2.0 
